refactor: move pmtree's sled wrapper into rln module and cleanup error type
