+++ rln/src/prelude.rs
@@ -7,7 +7,7 @@ pub use crate::circuit::{
     default_graph_multi, default_graph_single, default_zkey_multi, default_zkey_single,
 };
 #[cfg(not(target_arch = "wasm32"))]
-pub use crate::pm_tree::{FrOf, PmTree, PmTreeConfig, PmTreeProof};
+pub use crate::pm_tree::{FrOf, PmTree, PmTreeConfig, PmTreeMode, PmTreeProof};
 pub use crate::{
     circuit::{
         graph_from_raw, zkey_from_raw, ArkGroth16Backend, Curve, Fq, Fq2, Fr, G1Affine,
