+++ utils/Cargo.toml
@@ -14,8 +14,6 @@ bench = false
 [dependencies]
 ark-ff = { version = "0.5.0", default-features = false }
 num-bigint = { version = "0.4.6", default-features = false }
-pmtree = { package = "vacp2p_pmtree", version = "2.0.3", optional = true }
-sled = { version = "0.34.7", features = ["compression"], optional = true }
 serde_json = "1.0.150"
 rayon = "1.12.0"
 thiserror = "2.0.18"
@@ -31,7 +29,6 @@ criterion = { version = "0.8.2", features = ["html_reports"] }
 [features]
 default = []
 parallel = ["ark-ff/parallel"]
-pmtree-ft = ["pmtree", "sled"]
 
 [[bench]]
 name = "merkle_tree_benchmark"
