+++ utils/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod merkle_tree;
-pub mod pm_tree;
 pub mod poseidon;
