+++ utils/src/merkle_tree/error.rs
@@ -21,20 +21,19 @@ pub enum ZerokitMerkleTreeError {
     ComputingProofError,
     #[error("Invalid merkle proof length (!= tree depth)")]
     InvalidMerkleProof,
-    #[cfg(feature = "pmtree-ft")]
-    #[error("Pmtree error: {0}")]
-    PmtreeErrorKind(#[from] pmtree::PmtreeErrorKind),
+    #[error("Tree storage backend error: {0}")]
+    StorageBackend(String),
 }
 
 /// Errors that can occur while creating Merkle tree from config
 #[derive(Debug, thiserror::Error)]
 pub enum FromConfigError {
-    #[error("Error while reading pmtree config: {0}")]
+    #[error("Error while reading tree config: {0}")]
     JsonError(#[from] serde_json::Error),
-    #[error("Error while creating pmtree config: missing path")]
+    #[error("Error while creating tree config: missing path")]
     MissingPath,
-    #[error("Error while creating pmtree config: path already exists")]
+    #[error("Error while creating tree config: path already exists")]
     PathExists,
-    #[error("Error while creating pmtree default temp path: {0}")]
+    #[error("Error while creating tree default temp path: {0}")]
     IoError(#[from] std::io::Error),
 }
